Patches in support for more organisms with hyperGTest.
This just fixes some bugs that occured when certain organisms were
used with hyperGTest.  Now the code should be more robust to packages
having peculiar names etc. since the code now checks what type the
package is by looking at the database.  Also a bug was fixed that
caused things to depend on there being a CHR mapping.  This was a
problem with organisms that have only one chromosome such as E coli


git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/branches/RELEASE_2_3/madman/Rpacks/Category@35159 bc3139a8-67e5-0310-9ffc-ced21a209358